requirements.txt
streamlit
pandas
numpy
xgboost
scikit-learn
yfinance
ta
plotly
google-generativeai
python-dotenv
joblib